Added displayContact and contactByName functions
displayContact shows all 5 attributes of the contact; contactByName takes in the full name of the contact as a parameter and returns a pointer to that Contact (first name and last name can be jammed together like they would be if you got them from a trie). I also made the add function take input for all 5 data fields of Contact.