Made Stuff for In Ji
createContact method overloaded so you just pass in all the information.
Accessor method for getting the root of the tree you are currently sorted by.
Accessor methods for changing which tree you are pointing to. Always update the value of the tree head in main by calling treeHead() after changing which tree you want to use.
Accessor method "getSearchResults" that returns a vector of treeNodes. The contact information is stored in treeNode->c