Part 2/2 Overloaded Search
Suggestion on your part: Call the search function, then check searchResults. If the size of searchResults is 0, suggest to the user that they can input part of a first or last name or birthday but not a full legal name.

I will also make deleteFromAllTrees(Contact* c) and a postEdit(Contact* c) functions that will change the placement of a contact in all trees after you edit or delete the contact. These functions are not implemented yet.