Set up to work with postEdit() when I make it
I changed it so we set firstTreePointer, lastTreePointer, and birthTreePointer on a Contact during the add function instead of createContact. This will make it so a contact's pointers to its various treeNodes will update each time the contact is added to a tree (this is important since we will delete a treeNode and add the contact back to the tree in a new treeNode every time an edit is made to a contact).