Fixed postEdit function (untested)
postEdit will now update the values of firstTreePointer, lastTreePointer, and birthTreePointer (the variables which track the treeNodes in each tree that contain the Contact).

You can get rid of the editContact function if you want, or I can overload it with the same parameters that createContact has. If you decide to delete editContact and make your own edit function, just make sure to call postEdit afterward.